Formatting.

+++ test/libsolidity/semanticTests/end-to-end/code_access_create.sol
@@ -1,19 +1,26 @@
-		contract D {
-			uint x;
-			constructor() public { x = 7; }
-			function f() public view returns (uint) { return x; }
-		}
-		contract C {
-			function test() public returns (uint) {
-				bytes memory c = type(D).creationCode;
-				D d;
-				assembly {
-					d := create(0, add(c, 0x20), mload(c))
-				}
-				return d.f();
-			}
-		}
-	
+contract D {
+    uint256 x;
+
+    constructor() public {
+        x = 7;
+    }
+
+    function f() public view returns (uint256) {
+        return x;
+    }
+}
+
+
+contract C {
+    function test() public returns (uint256) {
+        bytes memory c = type(D).creationCode;
+        D d;
+        assembly {
+            d := create(0, add(c, 0x20), mload(c))
+        }
+        return d.f();
+    }
+}
+
 // ----
 // test() -> 7
-
